requirements/base.txt
Django==5.0.3
django-ckeditor==6.7.0
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
openai==0.28.0
pillow==10.2.0
requests==2.31.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-nested-inline==0.4.6
python-dotenv==1.0.1
drf-yasg==1.21.7
django-redis==5.4.0
whitenoise==6.6.0
django-storages==1.14.3
boto3==1.34.105
PaycomUz==3.0